nav bar updates
