+++ client/src/components/Signup.js
@@ -1,8 +1,7 @@
 import React from "react";
-import { useAppContext } from "../libs/contextLib";
+// import { useAppContext } from "../libs/contextLib";
 
 function Signup() {
-  const { userHasAuthenticated } = useAppContext();
 
   const signupFormHandler = async (event) => {
     event.preventDefault();
@@ -20,7 +19,7 @@ function Signup() {
   
       if (response.ok) {
         document.location.replace('/');
-        userHasAuthenticated(true);
+        sessionStorage.setItem("loggedIn",true);
       } else {
         alert(response.statusText);
       }
